examples/upload_result_sample/requirements.txt
--extra-index-url https://pypi.ngc.nvidia.com
nvidia-pytriton==0.5.0
numpy==1.26.3
requests==2.31.0
pillow==10.2.0
git+https://github.com/NVIDIA/nv-cloud-function-helpers.git@dce9d47e933fe12b381f718b86697dd7aece6b86
